Audio Pre-processing and Model Building
In this commit, the audio file was converted from stereo to mono and then train-val split was done. Then the audio was processed and converted to waveform, then to spectrogram and Sequential Model was trained. Then a test data was given which could classify which species of Cuckoo was calling.